requirements.txt
certifi==2021.5.30
MouseInfo==0.1.3
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.3
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PySide6==6.1.2
PyTweening==1.0.3
pywin32==301
shiboken6==6.1.2
six==1.16.0
wincertstore==0.2
